% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{decorate_sample_panoply}
\alias{decorate_sample_panoply}
\title{decorate_sample_panoply}
\usage{
decorate_sample_panoply(
  fcs,
  panoply_obj,
  superclus = NULL,
  colorscale = TRUE,
  superscale = FALSE,
  isLabeled = FALSE,
  cex = 0.5,
  ...
)
}
\arguments{
\item{fcs}{The data (either a flowFrame or a flowSet)}

\item{panoply_obj}{An object of type "panoply", the result of running panoply()}

\item{superclus}{TBD}

\item{colorscale}{logical.  Whether or not to draw the color wedge (default = TRUE)}

\item{superscale}{TBD}

\item{isLabeled}{TBD}

\item{cex}{Size of dots in peripheral panels.  Default 0.5}

\item{...}{Additional graphical parameters (see par)}
}
\description{
This function makes a nice picture of the result of panoply.

A picture of the result of running panoply().
}
\examples{
decorate_sample_panoply(fs_young, pan)
}
