% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{map_functional_names}
\alias{map_functional_names}
\title{title map_functional_names}
\usage{
map_functional_names(panoply_obj, defs_file)
}
\arguments{
\item{panoply_obj}{The result of running panoply()}

\item{defs_file}{A file containing the functional definitions.}
}
\value{
A decorated panoply object
}
\description{
Based on a user-specified table, assign symbolic names to clusters
}
