% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{panoply}
\alias{panoply}
\title{panoply}
\usage{
panoply(
  fcs,
  parameters = NULL,
  nRecursions = 12,
  perplexity = 40,
  nclust = NULL
)
}
\arguments{
\item{fcs}{The data (either a flowFrame or a flowSet)}

\item{parameters}{The parameters in fcs used for analysis}

\item{nRecursions}{The number of recursions in calculating the fingerprint (default = 12)}

\item{perplexity}{The perplexity value used for the imbedding of the bins (default = 40)}

\item{nclust}{The number of clusters you want panoply to make}
}
\value{
An object of class panoply, with the following elements:
\describe{
\item{mod}{The flowFPModel generated}
\item{mfi}{A list representation of the bin centers}
\item{centers}{The t-SNE map.  Dots represent bins}
\item{map}{The t-SNE map.  Dots represent bins}
\item{clustering}{A Vector containing cluster membership of the bins}
}
}
\description{
This function wraps most of what's needed.  It computes a fingerprint
model, calculates the multivariate bin centers, imbeds them in a t-SNE map, and
performs clustering in t-SNE space.

Panoply implements a workflow of doing Cytometric Fingerprint (CF) binning of
a flowFrame (or flowSet)
using flowFP to a relatively high resolution (default nRecursions of 12 results
in 4096 bins).  The bin centroids are then computed, using the median value of all
of the events in the bin for each of the included parameters.  Next, these
multivariate bin centroids are imbedded in a 2-dimensional t-SNE map. In this case,
the dots in the t-SNE map correspond to CF bins, NOT individual events.  Finally,
The bins are clustered using aglommerative hierarchical clustering in the 2
t-SNE dimensions, with a user-specified number of clusters.
The result is an object of class "panoply" that contains slots for all of
the important data elements.
}
\examples{
panoply(fs_young, parameters = c("CD4", "CD8", "CD45RA", "CCR7"), nclust = 15)
}
