% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{panoply_map_sample}
\alias{panoply_map_sample}
\title{Map a Sample to a Panoply Model}
\usage{
panoply_map_sample(ff, panoply_obj)
}
\arguments{
\item{ff}{A sample flowFrame}

\item{panoply_obj}{An object of type "panoply", the result of running panoply()}
}
\value{
Per-cluster counts and fractions
}
\description{
This function determines, for a single sample, the number of cells in each cluster.
}
