% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{panoply_phenobars}
\alias{panoply_phenobars}
\title{Visualize Cluster Phenotypes}
\usage{
panoply_phenobars(
  panoply_obj,
  parameters = colnames(panoply_obj$centers),
  cluster = 1,
  main = paste("Cluster", cluster)
)
}
\arguments{
\item{panoply_obj}{An object of type "panoply", the result of running panoply()}

\item{parameters}{Which parameters to include in the plot (default = all parameters)}

\item{cluster}{Which cluster to plot.}
}
\description{
Draw a "phenobar" representation of a cluster phenotype.  Bars have
a height equal to the medial value of the parameter and are
color-coded.  Error flags represent first and third quartiles of the bin centers
belonging to the cluster.
}
