% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panoply_methods.R
\name{plot_panoply_spread}
\alias{plot_panoply_spread}
\title{plot_panoply_spread}
\usage{
plot_panoply_spread(
  panoply,
  markers = NULL,
  mode = c("arithmetic", "robust"),
  cex = 50,
  proportional = TRUE,
  emph = TRUE,
  cex.lab = 2,
  highlight_clusters = NULL,
  contours = FALSE,
  show_cluster_number = NULL
)
}
\arguments{
\item{panoply}{The result of running panoply}

\item{markers}{Markers to include in the spread}

\item{mode}{Compute colors using either arithmetic or robust average}

\item{cex}{Scale factor for node size}

\item{proportional}{Logical.  Scale by the number of events in the cluster}

\item{emph}{Logical.  Emphasize each blob with a black line.}

\item{cex.lab}{Scale factor for titles of the individual markers}

\item{highlight_clusters}{Numeric vector indicating which clusters to highlight.}

\item{contours}{Logical. Plot the original t-SNE contours.}

\item{show_cluster_number}{If not null, print cluster numbers at the size given.}
}
\value{
N/A.
}
\description{
Draw a picture of the result of panoply for a given set of markers.
}
\examples{
plot_panoply_spread(panoply_obj)
}
