% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampled_flowset_young}
\alias{sampled_flowset_young}
\title{A sampled flowSet}
\format{A flowSet with 43 flowFrames.}
\source{
\url{https://flowrepository.org/id/FR-FCM-ZZGS}
}
\usage{
sampled_flowset_young
}
\description{
This flowSet originated from data in
\url{https://flowrepository.org/id/FR-FCM-ZZGS}.
The data were first compensated and then transformed using a custom biexponential transform.
They were then gated on singlet live events.  Then, each flowFrame in the
flowSet was sampled to 10,000 events.  Finally, only young (age <= 35 yrs)
was retained.
All this was done to reduce the size of the dataset for demonstration purposes.
}
\details{
A sampled flowSet
}
\keyword{datasets}
